refactor: change actions file structure and refactor code
