+++ src/app/registry.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useServerInsertedHTML } from 'next/navigation';
-import { Fragment, ReactNode, useState } from 'react';
+import { type ReactNode, useState } from 'react';
 import { ServerStyleSheet, StyleSheetManager } from 'styled-components';
 
 function StyledComponentsRegistry({ children }: { children: ReactNode }) {
@@ -13,11 +13,11 @@ function StyledComponentsRegistry({ children }: { children: ReactNode }) {
     const styles = styledComponentsStyleSheet.getStyleElement();
     styledComponentsStyleSheet.instance.clearTag();
 
-    return <Fragment>{styles}</Fragment>;
+    return <>{styles}</>;
   });
 
   if (typeof window !== 'undefined') {
-    return <Fragment>{children}</Fragment>;
+    return <>{children}</>;
   }
 
   return (
