+++ src/hooks/useMessages.ts
@@ -1,28 +1,30 @@
 import { useState } from 'react';
 import { v4 as uuidv4 } from 'uuid';
 
-import { createHumanMessage } from '@/app/actions';
 import useMainContext from './useMainContext';
+import { createHumanMessage } from '@/actions/messages';
 import { Message, Support } from '@/utils/definitions';
 
 function useMessages(data: Support) {
   const { user } = useMainContext();
-  const [isLoading, setIsLoading] = useState(false);
-  const [messages, setMessages] = useState([...data.messages]);
+  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const [messages, setMessages] = useState<Message[]>([...data.messages]);
 
   const addNewMessage = async (content: string) => {
     try {
       setIsLoading(true);
+
       const newMessage: Message = {
         id: uuidv4(),
-        role: 'collaborator',
+        conversation_id: data.conversation_id,
         content,
-        owner_profile: user,
         created_at: new Date().toISOString(),
+        role: 'collaborator',
+        owner_profile: user,
       };
 
       setMessages(messages.concat(newMessage));
-      await createHumanMessage(data.conversation_id, newMessage);
+      await createHumanMessage(newMessage);
     } catch (error) {
       setMessages(messages.slice(0, -1));
     } finally {
