+++ src/middleware.ts
@@ -1,6 +1,6 @@
 import { type NextRequest, NextResponse } from 'next/server';
 
-import { fetchProfile } from '@/app/actions';
+import { fetchProfile } from '@/actions/auth';
 import { updateSession } from '@/utils/supabase/middleware';
 
 export async function middleware(request: NextRequest) {
